Made some changes to the code to see what will improve delay.
ScanView now has a recycle event and expire event.

ScanViews will be destroyed after they are recycled n times in order to eventually cleanup unused objects in the recycle list.

The scan's base there time on when SetScan was called rather than when the scan was read in VisController's callback function.

The sorted list now sorts based on sequence of scans based on the seq value in the msg header.