Note: TfVisualizer was modified:
tree was set to public so LaserVisController could access it.

LaserScanView and LaserVisController work for the most part, decay times greater than 1 tend to get a bit laggy so an alternate method of showing points may need to be sought out.

What needs to be added at this point is a tilt to the points so everything isn't on one plane.