Added a RobotSubscriptionManagerUI
The RobotSubscriptionManagerUI is a custom Inspector for the RobotSubscriptionManager that gives it controls for all subscribed topics for all agents that are set up through the RobotSubscriptionManager. The RobotSubscriptionManager was modified to give the UI access to the necessary objects to handle modifying their settings.
The OdometryViewController was modified to change the scale of all present arrows instead of just new ones.