Modified SensorTFInterface, colors added to arrows and points
SensorTFInterface uses a virtual callBack function. All derived classes must call base.Start() but base.Update() is optional since not all classes using the interface need to have their transforms manipulated.
SensorTFInterface now has a MakeChildOfTF bool that will set the transform.parent equal to the TF that the sensor is associated with (frame_id in topic) when base.Update() is called. If set to false the base.Update() will set the transforms rotation and position to equal the TF instead. This can be set with the base constructor.

Colors have been added for arrows for OdometryViewController and points in LaserViewController , RobotSubscriptionManagerUI will need to be modified to handle color types.