LoadMesh working
TODO: add Materials
Started creating a way of maintaining a hierarchy of components and positioning them where they ought to be. This is kind of a pain and mostly unnecessary since the model can't be see until runtime where the tf's of each component should then be available. Leaving the code in, commented out in case it ends up being needed.