LoadMesh works more better
Currently works for the mrm_cover, Some hackey lines were added to make it work with Baxter. Whether or not this will work well on all meshes is unknown. Some playing with the rotations will be necessary to find a generic fix for all meshes.
TODO: 	figure out what's up with Baxters screen.
	Add material handling for locally defined materials.