Added potential orientation correction
In the Coladda files a up_axis field is set in assets. Based on what this is the mesh needs to be flipped on 1 or more axises. At the moment the script works on PR2, Valkyrie and most of Baxter. Baxter's screen is still broken.
I also removed material setting for meshes, may look into the Coladda files settings to properly set material if necessary since it will provide more info then the urdf. At the moment the the Collada files tend to have their proper materials set by default.