Added a script that wraps the CompressedImageDisplay script.
RobotCameraView takes CompressedImageDisplay and applies it to a plane and transforms the plane based on a given odom topic. As it stands the frame ID for the compressed image topic only seems to return the namespace of it's associated robot which is not helpful in generically assigning it's position. Explicitly stating an associated odom topic seems like a decent enough fix.
Also modified some prefabs and added a RobotCameraView prefab.

 NOTE: laserScanView is acting funky. NaN's were being passed to the position of the spheres, crashing the program and the points seem to get displaced quite  a bit when the robot is rotating.